clinic-cli/requirements.txt
google-genai
pydantic
pyaudio
python-dotenv